Remove test utils crate / remove create_test xtask
