+++ Cargo.toml
@@ -3,7 +3,6 @@ resolver = "2"
 members = [
   "crates/cairo-lint-cli",
   "crates/cairo-lint-core",
-  "crates/cairo-lint-test-utils",
   "xtask"
 ]
 
@@ -39,7 +38,6 @@ cairo-lang-semantic = "*"
 cairo-lang-starknet = "*"
 cairo-lang-syntax = "*"
 cairo-lang-test-plugin = "*"
-cairo-lang-test-utils = "*"
 cairo-lang-utils = "*"
 clap = { version = "4.5.27", features = ["derive"]}
 ctor = "0.2.9"
@@ -83,5 +81,4 @@ cairo-lang-starknet-classes = { git = "https://github.com/starkware-libs/cairo",
 cairo-lang-syntax = { git = "https://github.com/starkware-libs/cairo", rev = "b3bb598ff815e8e2b2d8418c9bb916578873805e" }
 cairo-lang-syntax-codegen = { git = "https://github.com/starkware-libs/cairo", rev = "b3bb598ff815e8e2b2d8418c9bb916578873805e" }
 cairo-lang-test-plugin = { git = "https://github.com/starkware-libs/cairo", rev = "b3bb598ff815e8e2b2d8418c9bb916578873805e" }
-cairo-lang-test-utils = { git = "https://github.com/starkware-libs/cairo", rev = "b3bb598ff815e8e2b2d8418c9bb916578873805e" }
 cairo-lang-utils = { git = "https://github.com/starkware-libs/cairo", rev = "b3bb598ff815e8e2b2d8418c9bb916578873805e" }
