--- crates/cairo-lint-test-utils/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "cairo-lint-test-utils"
-version.workspace = true
-edition.workspace = true
-repository.workspace = true
-license-file.workspace = true
-
-[dependencies]
-cairo-lang-utils.workspace = true
-cairo-lang-test-utils.workspace = true
-cairo-lang-filesystem.workspace = true
-cairo-lang-diagnostics.workspace = true
-cairo-lang-semantic.workspace = true
-cairo-lang-defs.workspace = true
-cairo-lang-compiler.workspace = true
