+++ xtask/src/main.rs
@@ -21,7 +21,7 @@ macro_rules! command {
   }
 }
 
-command!(Command(upgrade, create_test, sync_version,));
+command!(Command(upgrade, sync_version,));
 
 #[derive(Parser)]
 struct Args {
